backend/requirements.txt
kafka-python-ng>=2.2.3
fastapi==0.111.0
uvicorn==0.30.1
sse-starlette==2.1.0
python-dotenv==1.0.0
